feat: move reset in the models page to the top menue
