+++ src/components/ModelsResetDialog/styles.ts
@@ -1,12 +1,10 @@
 import {StyleSheet} from 'react-native';
 
 export const styles = StyleSheet.create({
-  fab: {
-    bottom: 0,
-    right: 16,
+  paragraph: {
+    marginBottom: 8,
   },
-  icon: {
-    width: 24,
-    height: 24,
+  bold: {
+    fontWeight: 'bold',
   },
 });
